Don't restart gdm for installed tests

+++ features/steps/steps.py
@@ -1,13 +1,15 @@
 from behave import step
 from time import sleep
-from datetime import datetime
 import subprocess
 import os
 
 
 @step(u'Start {service_name} service')
-def start_systemd_service(context, service_name):
-    cmd = "sudo systemctl stop %s && sudo systemctl start %s" % (service_name, service_name)
+@step(u'Start {service_name} service (without {restart} it)')
+def start_systemd_service(context, service_name, restart=False):
+    cmd = "sudo systemctl start %s" % service_name
+    if restart:
+        cmd = "sudo systemctl stop %s && sudo systemctl start %s" % (service_name, service_name)
     try:
         print("Running '%s'" % cmd)
         subprocess.check_output(cmd, shell=True)
